Add water color form field, its display in view report screen, fixed 2 bugs in view report screen related to coordinates and deleting null imagen, and added compatiblity for adminstatus arguments for the bottom list

+++ lib/widgets/view_report/view_report_data.dart
@@ -9,7 +9,7 @@ import 'package:ocean_change/widgets/view_report/view_screen_args.dart';
 class ViewReportData extends StatefulWidget {
   final ViewScreenArgs viewScreenArgs;
 
-  ViewReportData({super.key, required this.viewScreenArgs});
+  const ViewReportData({super.key, required this.viewScreenArgs});
 
   @override
   State<ViewReportData> createState() => _ViewReportDataState();
@@ -66,12 +66,21 @@ class _ViewReportDataState extends State<ViewReportData> {
       ),
       loadWaterTemp(widget.viewScreenArgs.userReport.waterTemp),
       Padding(
+          padding: const EdgeInsets.fromLTRB(8.0, 4.0, 8.0, 4.0),
+          child: Row(
+            children: [
+              const Text('Water Color: ', style: Styles.viewScreenSmallLabels),
+              Text(widget.viewScreenArgs.userReport.waterColor ?? 'none',
+                  style: Styles.viewScreenSmallData),
+            ],
+          )),
+      Padding(
         padding: const EdgeInsets.fromLTRB(8.0, 4.0, 8.0, 4.0),
         child: Row(
           children: [
             const Text('Location:', style: Styles.viewScreenSmallLabels),
             Text(
-              ' ${widget.viewScreenArgs.userReport.geopoint?.latitude.toString().substring(0, 8)}° N, ${widget.viewScreenArgs.userReport.geopoint?.longitude.toString().substring(0, 10)}° W',
+              ' ${double.parse(widget.viewScreenArgs.userReport.geopoint!.latitude.toStringAsFixed(2))}° N, ${double.parse(widget.viewScreenArgs.userReport.geopoint!.longitude.toStringAsFixed(4))}° W',
               style: Styles.viewScreenSmallData,
             ),
           ],
@@ -123,8 +132,14 @@ class _ViewReportDataState extends State<ViewReportData> {
         .collection("reports")
         .doc(userReport.id)
         .delete();
-    final imageRef = FirebaseStorage.instance.refFromURL(userReport.photoURL.toString());
-    await imageRef.delete();
-    setState(() {Navigator.pop(context);});
+
+    if (userReport.photoURL != null) {
+      final imageRef =
+          FirebaseStorage.instance.refFromURL(userReport.photoURL.toString());
+      await imageRef.delete();
+    }
+    setState(() {
+      Navigator.pop(context);
+    });
   }
 }
